Rename: listings to usagePaths
