+++ src/main/resources/admin/tools/part-finder/usagePaths.ts
@@ -87,9 +87,18 @@ export function getUsagePaths(
 
     const componentData = (component || {})[targetType] || {};
     const componentDescriptor = componentData.descriptor;
+
+    /*log.info(
+      "\n---\n" +
+        JSON.stringify({ component, componentData, componentDescriptor, targetDescriptor, targetType }, null, 2),
+    );*/
+
     if (component?.type === targetType && componentDescriptor === targetDescriptor) {
       const splitDescriptor = componentDescriptor.split(":");
       const componentConfig = ((componentData.config || {})[splitDescriptor[0]] || {})[splitDescriptor[1]];
+
+      //log.info("\n!!!\n" + JSON.stringify({ splitDescriptor, componentConfig }, null, 2));
+
       pushUsagePath(component.path, usagePaths, componentConfig, getvalueParam);
     }
   });
